src/visualizer/requirements.txt
azure-cosmos>=4.8.0
python-dotenv>=1.0.1
matplotlib>=3.7.0
